click-sample/requirements.txt
--index https://mirrors.aliyun.com/pypi/simple
click>=7.1.0
python-dotenv>=0.15.0
pytest>=6.1.0
